refactor: PostEditModal에 내려주는 props 일부 제거
